requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
itsdangerous
pytest
httpx
pillow
